requirements_orin.txt
tensorflow==2.16.1+nv24.07
sionna-no-rt < 2.0
numpy
matplotlib
nanobind
onnx
tf2onnx
pytest
#tensorrt
pyzmq
pyyaml

